Implement various time changes which usually break software, e.g. start and end of the month, end of the year, leap days.
This implements the basic functionality however, the timezone changes are still going wrong.
What still needs to be implemented is some kind of method which finds
the appropiate time and day we wanted to select.

What happens now is that since the newly created date is in a different timezone the the is
shifted backwards or forwards one hour (or more, depending on the local timezone).

Which causes, for example, the 27th of february to be selected when the last day of the month
mode is used, instead of the 28, which is what we wanted.